testing_sys/requirements.txt
Babel==2.6.0
Django==2.2
django-model-utils==3.1.2
django-phonenumber-field==2.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.2
phonenumbers==8.10.10
Pillow==6.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
